+++ test/unit/spec/createlocaltracks.ts
@@ -2,8 +2,8 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
+import { FakeMediaStreamTrack, fakeGetUserMedia  } from '../../lib/fakemediastream';
 import { createLocalTracks } from '../../../lib/createlocaltracks';
-import { fakeGetUserMedia, FakeMediaStreamTrack } from '../../lib/fakemediastream';
 
 describe('createLocalTracks', () => {
   [